requirements.txt
streamlit
pandas
plotly
requests
reportlab
python-dotenv